refactor(commands): extract /permissions command into permissions.rs
